requirements.txt
flask
flask-login
flask-wtf
flask-sqlalchemy
WTForms-SQLAlchemy